requirements.txt
# FinSight Dependencies

# Core
asyncpg==0.29.0
pandas==2.2.0
numpy==1.26.3

# Data Sources
yfinance==0.2.36
requests==2.31.0
aiohttp==3.9.1
feedparser==6.0.10
beautifulsoup4==4.12.3

# ML
scikit-learn==1.4.0

# Dashboard (legacy)
streamlit==1.31.0
plotly==5.18.0

# Web API (NEW)
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
email-validator==2.1.0
slowapi==0.1.9

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
